Refactor sources
