+++ hellokoin/src/main/java/com/airbnb/mvrx/hellokoin/di/Modules.kt
@@ -1,6 +1,10 @@
 package com.airbnb.mvrx.hellokoin.di
 
 import com.airbnb.mvrx.hellokoin.*
+import com.airbnb.mvrx.hellokoin.screens.hello.HelloState
+import com.airbnb.mvrx.hellokoin.screens.hello.HelloViewModel
+import com.airbnb.mvrx.hellokoin.screens.scopedhello.ScopedHelloState
+import com.airbnb.mvrx.hellokoin.screens.scopedhello.ScopedHelloViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
