+++ hellokoin/src/main/res/navigation/main_navigation.xml
@@ -19,12 +19,12 @@
 
     <fragment
         android:id="@+id/helloFragment"
-        android:name="com.airbnb.mvrx.hellokoin.HelloFragment"
+        android:name="com.airbnb.mvrx.hellokoin.screens.hello.HelloFragment"
         android:label="HelloFragment" />
 
     <fragment
         android:id="@+id/scopedHelloFragment"
-        android:name="com.airbnb.mvrx.hellokoin.ScopedHelloFragment"
+        android:name="com.airbnb.mvrx.hellokoin.screens.scopedhello.ScopedHelloFragment"
         android:label="ScopedHelloFragment" />
 
 </navigation>
