+++ hellokoin/src/test/java/com/airbnb/mvrx/hellokoin/HelloViewModelTest.kt
@@ -1,5 +1,7 @@
 package com.airbnb.mvrx.hellokoin
 
+import com.airbnb.mvrx.hellokoin.screens.hello.HelloState
+import com.airbnb.mvrx.hellokoin.screens.hello.HelloViewModel
 import com.airbnb.mvrx.test.MvRxTestRule
 import com.airbnb.mvrx.withState
 import io.mockk.every
